feat: mocks/spy | afterEach | props test.
A spy was created to test a ItemCounter mock along with its props.

afterEach was used to clean all of the tests, avoiding false negatives referred to quantity of the component renders.

A prop test was created to make sure how many times the component was called (that's why we used afterEach) and that the component was supposed to be called including its properties.